cli/pyproject.toml
[project]
name = "kcontext-cli"
version = "0.1.0"
description = "kcontext data pipeline CLI - fetch, build, push YouTube subtitle data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.15.0",
    "yt-dlp>=2024.0.0",
    "youtube-transcript-api>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
kcontext = "kcontext_cli.main:app"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.ruff.format]
quote-style = "double"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.15.4",
    "pytest>=8.0",
    "pytest-mock>=3.14",
]
